+++ lib/promscrape/discovery/kubernetes/api_watcher_test.go
@@ -129,7 +129,7 @@ func TestGetAPIPathsWithNamespaces(t *testing.T) {
 	})
 
 	// role=endpointslices
-	f("endpointslices", nil, nil, []string{"/apis/discovery.k8s.io/v1beta1/endpointslices"})
+	f("endpointslices", nil, nil, []string{"/apis/discovery.k8s.io/v1/endpointslices"})
 	f("endpointslices", []string{"x", "y"}, []Selector{
 		{
 			Role:  "endpointslices",
@@ -137,12 +137,12 @@ func TestGetAPIPathsWithNamespaces(t *testing.T) {
 			Label: "label",
 		},
 	}, []string{
-		"/apis/discovery.k8s.io/v1beta1/namespaces/x/endpointslices?labelSelector=label&fieldSelector=field",
-		"/apis/discovery.k8s.io/v1beta1/namespaces/y/endpointslices?labelSelector=label&fieldSelector=field",
+		"/apis/discovery.k8s.io/v1/namespaces/x/endpointslices?labelSelector=label&fieldSelector=field",
+		"/apis/discovery.k8s.io/v1/namespaces/y/endpointslices?labelSelector=label&fieldSelector=field",
 	})
 
 	// role=ingress
-	f("ingress", nil, nil, []string{"/apis/networking.k8s.io/v1beta1/ingresses"})
+	f("ingress", nil, nil, []string{"/apis/networking.k8s.io/v1/ingresses"})
 	f("ingress", []string{"x", "y"}, []Selector{
 		{
 			Role:  "node",
@@ -161,8 +161,8 @@ func TestGetAPIPathsWithNamespaces(t *testing.T) {
 			Label: "baaa",
 		},
 	}, []string{
-		"/apis/networking.k8s.io/v1beta1/namespaces/x/ingresses?labelSelector=cde%2Cbaaa&fieldSelector=abc",
-		"/apis/networking.k8s.io/v1beta1/namespaces/y/ingresses?labelSelector=cde%2Cbaaa&fieldSelector=abc",
+		"/apis/networking.k8s.io/v1/namespaces/x/ingresses?labelSelector=cde%2Cbaaa&fieldSelector=abc",
+		"/apis/networking.k8s.io/v1/namespaces/y/ingresses?labelSelector=cde%2Cbaaa&fieldSelector=abc",
 	})
 }
 
@@ -577,9 +577,9 @@ func TestGetScrapeWorkObjects(t *testing.T) {
 			initAPIObjectsByRole: map[string][]byte{
 				"ingress": []byte(`{
   "kind": "IngressList",
-  "apiVersion": "extensions/v1beta1",
+  "apiVersion": "extensions/v1",
   "metadata": {
-    "selfLink": "/apis/extensions/v1beta1/ingresses",
+    "selfLink": "/apis/extensions/v1/ingresses",
     "resourceVersion": "351452"
   },
   "items": [
@@ -678,9 +678,9 @@ func TestGetScrapeWorkObjects(t *testing.T) {
 			initAPIObjectsByRole: map[string][]byte{
 				"endpointslices": []byte(`{
   "kind": "EndpointSliceList",
-  "apiVersion": "discovery.k8s.io/v1beta1",
+  "apiVersion": "discovery.k8s.io/v1",
   "metadata": {
-    "selfLink": "/apis/discovery.k8s.io/v1beta1/endpointslices",
+    "selfLink": "/apis/discovery.k8s.io/v1/endpointslices",
     "resourceVersion": "1177"
   },
   "items": [