+++ lib/promscrape/discovery/kubernetes/ingress_test.go
@@ -30,9 +30,9 @@ func TestParseIngressListSuccess(t *testing.T) {
 	data := `
 {
   "kind": "IngressList",
-  "apiVersion": "extensions/v1beta1",
+  "apiVersion": "extensions/v1",
   "metadata": {
-    "selfLink": "/apis/extensions/v1beta1/ingresses",
+    "selfLink": "/apis/extensions/v1/ingresses",
     "resourceVersion": "351452"
   },
   "items": [
@@ -40,13 +40,13 @@ func TestParseIngressListSuccess(t *testing.T) {
       "metadata": {
         "name": "test-ingress",
         "namespace": "default",
-        "selfLink": "/apis/extensions/v1beta1/namespaces/default/ingresses/test-ingress",
+        "selfLink": "/apis/extensions/v1/namespaces/default/ingresses/test-ingress",
         "uid": "6d3f38f9-de89-4bc9-b273-c8faf74e8a27",
         "resourceVersion": "351445",
         "generation": 1,
         "creationTimestamp": "2020-04-13T16:43:52Z",
         "annotations": {
-          "kubectl.kubernetes.io/last-applied-configuration": "{\"apiVersion\":\"networking.k8s.io/v1beta1\",\"kind\":\"Ingress\",\"metadata\":{\"annotations\":{},\"name\":\"test-ingress\",\"namespace\":\"default\"},\"spec\":{\"backend\":{\"serviceName\":\"testsvc\",\"servicePort\":80}}}\n"
+          "kubectl.kubernetes.io/last-applied-configuration": "{\"apiVersion\":\"networking.k8s.io/v1\",\"kind\":\"Ingress\",\"metadata\":{\"annotations\":{},\"name\":\"test-ingress\",\"namespace\":\"default\"},\"spec\":{\"backend\":{\"serviceName\":\"testsvc\",\"servicePort\":80}}}\n"
         }
       },
       "spec": {
@@ -85,7 +85,7 @@ func TestParseIngressListSuccess(t *testing.T) {
 	expectedLabelss := [][]prompbmarshal.Label{
 		discoveryutils.GetSortedLabels(map[string]string{
 			"__address__": "foobar",
-			"__meta_kubernetes_ingress_annotation_kubectl_kubernetes_io_last_applied_configuration":        `{"apiVersion":"networking.k8s.io/v1beta1","kind":"Ingress","metadata":{"annotations":{},"name":"test-ingress","namespace":"default"},"spec":{"backend":{"serviceName":"testsvc","servicePort":80}}}` + "\n",
+			"__meta_kubernetes_ingress_annotation_kubectl_kubernetes_io_last_applied_configuration":        `{"apiVersion":"networking.k8s.io/v1","kind":"Ingress","metadata":{"annotations":{},"name":"test-ingress","namespace":"default"},"spec":{"backend":{"serviceName":"testsvc","servicePort":80}}}` + "\n",
 			"__meta_kubernetes_ingress_annotationpresent_kubectl_kubernetes_io_last_applied_configuration": "true",
 			"__meta_kubernetes_ingress_host":   "foobar",
 			"__meta_kubernetes_ingress_name":   "test-ingress",